Moved componentWillMount to parent
